Emacs leader key on both sides
Having the Emacs leader (C-c) binding only on the right side
was putting stress on my fingers. For example, it's difficult
to do C-c b h switching from right to left. It would be much
easier to be able to do "C-c b" with the left hand.

This removes the earlier 'sticky' behavior for 'Ctrl' on the
left side.